Get TCP happy path working
