+++ crates/spin-test-virt/src/wasi.rs
@@ -5,6 +5,7 @@ mod filesystem;
 pub mod http;
 pub mod http_helper;
 pub mod io;
+mod tcp;
 
 use crate::bindings::exports::wasi;
 use crate::Component;
@@ -152,9 +153,17 @@ impl wasi::cli::exit::Guest for Component {
 
 impl wasi::sockets::instance_network::Guest for Component {
     fn instance_network() -> wasi::sockets::instance_network::Network {
-        todo!()
+        wasi::sockets::instance_network::Network::new(Network)
     }
 }
+
+impl wasi::sockets::network::Guest for Component {
+    type Network = Network;
+}
+
+pub struct Network;
+
+impl wasi::sockets::network::GuestNetwork for Network {}
 
 impl wasi::sockets::ip_name_lookup::Guest for Component {
     type ResolveAddressStream = ResolveAddressStream;
@@ -183,187 +192,6 @@ impl wasi::sockets::ip_name_lookup::GuestResolveAddressStream for ResolveAddress
     }
 
     fn subscribe(&self) -> wasi::sockets::ip_name_lookup::Pollable {
-        todo!()
-    }
-}
-
-impl wasi::sockets::network::Guest for Component {
-    type Network = Network;
-}
-
-pub struct Network;
-
-impl wasi::sockets::network::GuestNetwork for Network {}
-
-impl wasi::sockets::tcp::Guest for Component {
-    type TcpSocket = TcpSocket;
-}
-
-pub struct TcpSocket;
-
-impl wasi::sockets::tcp::GuestTcpSocket for TcpSocket {
-    fn start_bind(
-        &self,
-        network: wasi::sockets::tcp::NetworkBorrow<'_>,
-        local_address: wasi::sockets::tcp::IpSocketAddress,
-    ) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn finish_bind(&self) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn start_connect(
-        &self,
-        network: wasi::sockets::tcp::NetworkBorrow<'_>,
-        remote_address: wasi::sockets::tcp::IpSocketAddress,
-    ) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn finish_connect(
-        &self,
-    ) -> Result<
-        (
-            wasi::sockets::tcp::InputStream,
-            wasi::sockets::tcp::OutputStream,
-        ),
-        wasi::sockets::tcp::ErrorCode,
-    > {
-        todo!()
-    }
-
-    fn start_listen(&self) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn finish_listen(&self) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn accept(
-        &self,
-    ) -> Result<
-        (
-            wasi::sockets::tcp::TcpSocket,
-            wasi::sockets::tcp::InputStream,
-            wasi::sockets::tcp::OutputStream,
-        ),
-        wasi::sockets::tcp::ErrorCode,
-    > {
-        todo!()
-    }
-
-    fn local_address(
-        &self,
-    ) -> Result<wasi::sockets::tcp::IpSocketAddress, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn remote_address(
-        &self,
-    ) -> Result<wasi::sockets::tcp::IpSocketAddress, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn is_listening(&self) -> bool {
-        todo!()
-    }
-
-    fn address_family(&self) -> wasi::sockets::tcp::IpAddressFamily {
-        todo!()
-    }
-
-    fn set_listen_backlog_size(&self, value: u64) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn keep_alive_enabled(&self) -> Result<bool, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_keep_alive_enabled(&self, value: bool) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn keep_alive_idle_time(
-        &self,
-    ) -> Result<wasi::sockets::tcp::Duration, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_keep_alive_idle_time(
-        &self,
-        value: wasi::sockets::tcp::Duration,
-    ) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn keep_alive_interval(
-        &self,
-    ) -> Result<wasi::sockets::tcp::Duration, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_keep_alive_interval(
-        &self,
-        value: wasi::sockets::tcp::Duration,
-    ) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn keep_alive_count(&self) -> Result<u32, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_keep_alive_count(&self, value: u32) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn hop_limit(&self) -> Result<u8, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_hop_limit(&self, value: u8) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn receive_buffer_size(&self) -> Result<u64, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_receive_buffer_size(&self, value: u64) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn send_buffer_size(&self) -> Result<u64, wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn set_send_buffer_size(&self, value: u64) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-
-    fn subscribe(&self) -> wasi::sockets::tcp::Pollable {
-        todo!()
-    }
-
-    fn shutdown(
-        &self,
-        shutdown_type: wasi::sockets::tcp::ShutdownType,
-    ) -> Result<(), wasi::sockets::tcp::ErrorCode> {
-        todo!()
-    }
-}
-
-impl wasi::sockets::tcp_create_socket::Guest for Component {
-    fn create_tcp_socket(
-        address_family: wasi::sockets::tcp_create_socket::IpAddressFamily,
-    ) -> Result<
-        wasi::sockets::tcp_create_socket::TcpSocket,
-        wasi::sockets::tcp_create_socket::ErrorCode,
-    > {
         todo!()
     }
 }
